Add files via upload
Watching misfits and magic on the dropout streaming service. It's a TTRPG, but for dice rolls, if you roll the max number on the dice, it "explodes" and you get another roll, so you can really get any number with any dice. Really trying to compare the probability of difference dices.